refactor: core.resource 패키지 변경 사항 반영
