+++ app/src/main/java/com/into/websoso/ui/activityDetail/ActivityDetailActivity.kt
@@ -21,14 +21,14 @@ import com.into.websoso.core.common.ui.model.ResultFrom.BlockUser
 import com.into.websoso.core.common.ui.model.ResultFrom.FeedDetailBack
 import com.into.websoso.core.common.ui.model.ResultFrom.FeedDetailRemoved
 import com.into.websoso.core.common.util.showWebsosoSnackBar
+import com.into.websoso.core.resource.R.drawable.ic_blocked_user_snack_bar
+import com.into.websoso.core.resource.R.string.block_user_success_message
+import com.into.websoso.core.resource.R.string.feed_removed_feed_snackbar
+import com.into.websoso.core.resource.R.string.my_activity_detail_title
+import com.into.websoso.core.resource.R.string.other_user_page_activity
 import com.into.websoso.databinding.ActivityActivityDetailBinding
 import com.into.websoso.databinding.MenuMyActivityPopupBinding
 import com.into.websoso.databinding.MenuOtherUserActivityPopupBinding
-import com.into.websoso.resource.R.drawable.ic_blocked_user_snack_bar
-import com.into.websoso.resource.R.string.block_user_success_message
-import com.into.websoso.resource.R.string.feed_removed_feed_snackbar
-import com.into.websoso.resource.R.string.my_activity_detail_title
-import com.into.websoso.resource.R.string.other_user_page_activity
 import com.into.websoso.ui.activityDetail.adapter.ActivityDetailAdapter
 import com.into.websoso.ui.createFeed.CreateFeedActivity
 import com.into.websoso.ui.feedDetail.FeedDetailActivity
