+++ app/src/main/java/com/into/websoso/ui/blockedUsers/BlockedUsersActivity.kt
@@ -8,9 +8,9 @@ import com.into.websoso.R.layout.activity_blocked_users
 import com.into.websoso.core.common.ui.base.BaseActivity
 import com.into.websoso.core.common.util.SingleEventHandler
 import com.into.websoso.core.common.util.showWebsosoSnackBar
+import com.into.websoso.core.resource.R.drawable.ic_blocked_user_snack_bar
+import com.into.websoso.core.resource.R.string.blocked_users_unblocked
 import com.into.websoso.databinding.ActivityBlockedUsersBinding
-import com.into.websoso.resource.R.drawable.ic_blocked_user_snack_bar
-import com.into.websoso.resource.R.string.blocked_users_unblocked
 import com.into.websoso.ui.blockedUsers.adapter.BlockedUsersAdapter
 import dagger.hilt.android.AndroidEntryPoint
