+++ app/src/main/java/com/into/websoso/ui/createFeed/CreatingFeedDialogFragment.kt
@@ -5,10 +5,10 @@ import android.view.View
 import com.into.websoso.R.layout.dialog_report_popup_menu
 import com.into.websoso.core.common.ui.base.BaseDialogFragment
 import com.into.websoso.core.common.util.SingleEventHandler
+import com.into.websoso.core.resource.R.string.remove_popup_menu_keep_creating
+import com.into.websoso.core.resource.R.string.remove_popup_menu_stop_creating
+import com.into.websoso.core.resource.R.string.tv_remove_popup_menu_stop_creating
 import com.into.websoso.databinding.DialogReportPopupMenuBinding
-import com.into.websoso.resource.R.string.remove_popup_menu_keep_creating
-import com.into.websoso.resource.R.string.remove_popup_menu_stop_creating
-import com.into.websoso.resource.R.string.tv_remove_popup_menu_stop_creating
 import com.into.websoso.ui.main.feed.FeedFragment.FeedDialogClickListener
 
 class CreatingFeedDialogFragment :
