+++ app/src/main/java/com/into/websoso/ui/detailExploreResult/DetailExploreResultKeywordFragment.kt
@@ -24,10 +24,10 @@ import com.into.websoso.core.common.ui.model.CategoriesModel.Companion.findKeywo
 import com.into.websoso.core.common.util.SingleEventHandler
 import com.into.websoso.core.common.util.toFloatPxFromDp
 import com.into.websoso.core.common.util.tracker.Tracker
+import com.into.websoso.core.resource.R.drawable.ic_novel_rating_keword_remove
+import com.into.websoso.core.resource.R.string.detail_explore_search_hint
+import com.into.websoso.core.resource.R.string.inquire_link
 import com.into.websoso.databinding.FragmentDetailExploreResultKeywordBinding
-import com.into.websoso.resource.R.drawable.ic_novel_rating_keword_remove
-import com.into.websoso.resource.R.string.detail_explore_search_hint
-import com.into.websoso.resource.R.string.inquire_link
 import com.into.websoso.ui.detailExplore.keyword.DetailExploreClickListener
 import com.into.websoso.ui.detailExplore.keyword.adapter.DetailExploreKeywordAdapter
 import com.into.websoso.ui.detailExploreResult.model.DetailExploreResultUiState
