+++ app/src/main/java/com/into/websoso/ui/feedDetail/FeedDetailActivity.kt
@@ -40,14 +40,14 @@ import com.into.websoso.core.common.util.showWebsosoSnackBar
 import com.into.websoso.core.common.util.toFloatPxFromDp
 import com.into.websoso.core.common.util.toIntPxFromDp
 import com.into.websoso.core.common.util.tracker.Tracker
+import com.into.websoso.core.resource.R.drawable.ic_blocked_user_snack_bar
+import com.into.websoso.core.resource.R.string.feed_popup_menu_content_isMyFeed
+import com.into.websoso.core.resource.R.string.feed_popup_menu_content_report_isNotMyFeed
+import com.into.websoso.core.resource.R.string.other_user_page_withdraw_user
 import com.into.websoso.databinding.ActivityFeedDetailBinding
 import com.into.websoso.databinding.DialogRemovePopupMenuBinding
 import com.into.websoso.databinding.DialogReportPopupMenuBinding
 import com.into.websoso.databinding.MenuFeedPopupBinding
-import com.into.websoso.resource.R.drawable.ic_blocked_user_snack_bar
-import com.into.websoso.resource.R.string.feed_popup_menu_content_isMyFeed
-import com.into.websoso.resource.R.string.feed_popup_menu_content_report_isNotMyFeed
-import com.into.websoso.resource.R.string.other_user_page_withdraw_user
 import com.into.websoso.ui.createFeed.CreateFeedActivity
 import com.into.websoso.ui.feedDetail.FeedDetailActivity.MenuType.COMMENT
 import com.into.websoso.ui.feedDetail.FeedDetailActivity.MenuType.FEED
