+++ app/src/main/java/com/into/websoso/ui/main/feed/dialog/FeedReportDialogFragment.kt
@@ -6,11 +6,11 @@ import com.into.websoso.R.layout.dialog_report_popup_menu
 import com.into.websoso.core.common.ui.base.BaseDialogFragment
 import com.into.websoso.core.common.util.SingleEventHandler
 import com.into.websoso.core.common.util.tracker.Tracker
+import com.into.websoso.core.resource.R.string.report_popup_menu_impertinence_comment
+import com.into.websoso.core.resource.R.string.report_popup_menu_impertinence_feed
+import com.into.websoso.core.resource.R.string.report_popup_menu_spoiling_comment
+import com.into.websoso.core.resource.R.string.report_popup_menu_spoiling_feed
 import com.into.websoso.databinding.DialogReportPopupMenuBinding
-import com.into.websoso.resource.R.string.report_popup_menu_impertinence_comment
-import com.into.websoso.resource.R.string.report_popup_menu_impertinence_feed
-import com.into.websoso.resource.R.string.report_popup_menu_spoiling_comment
-import com.into.websoso.resource.R.string.report_popup_menu_spoiling_feed
 import com.into.websoso.ui.main.feed.FeedFragment.FeedDialogClickListener
 import com.into.websoso.ui.main.feed.dialog.ReportMenuType.IMPERTINENCE_COMMENT
 import com.into.websoso.ui.main.feed.dialog.ReportMenuType.IMPERTINENCE_FEED
