+++ app/src/main/java/com/into/websoso/ui/main/home/HomeFragment.kt
@@ -21,9 +21,9 @@ import com.into.websoso.core.common.ui.model.ResultFrom.NovelDetailBack
 import com.into.websoso.core.common.ui.model.ResultFrom.ProfileEditSuccess
 import com.into.websoso.core.common.util.collectWithLifecycle
 import com.into.websoso.core.common.util.tracker.Tracker
+import com.into.websoso.core.resource.R.string.home_nickname_interest_feed
 import com.into.websoso.data.repository.AuthRepository
 import com.into.websoso.databinding.FragmentHomeBinding
-import com.into.websoso.resource.R.string.home_nickname_interest_feed
 import com.into.websoso.ui.feedDetail.FeedDetailActivity
 import com.into.websoso.ui.main.MainViewModel
 import com.into.websoso.ui.main.home.adpater.PopularFeedsAdapter
