+++ app/src/main/java/com/into/websoso/ui/notification/component/NotificationAppBar.kt
@@ -18,11 +18,11 @@ import androidx.compose.ui.text.style.TextAlign.Companion.Center
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.into.websoso.core.common.util.clickableWithoutRipple
+import com.into.websoso.core.resource.R.drawable.ic_notification_back
+import com.into.websoso.core.resource.R.string.notification_notification
 import com.into.websoso.designsystem.theme.Black
 import com.into.websoso.designsystem.theme.WebsosoTheme
 import com.into.websoso.designsystem.theme.White
-import com.into.websoso.resource.R.drawable.ic_notification_back
-import com.into.websoso.resource.R.string.notification_notification
 
 @Composable
 fun NotificationAppBar(
