+++ app/src/main/java/com/into/websoso/ui/novelDetail/NovelDetailActivity.kt
@@ -30,19 +30,19 @@ import com.into.websoso.core.common.util.showWebsosoSnackBar
 import com.into.websoso.core.common.util.toFloatPxFromDp
 import com.into.websoso.core.common.util.toIntPxFromDp
 import com.into.websoso.core.common.util.tracker.Tracker
+import com.into.websoso.core.resource.R.drawable.ic_novel_detail_check
+import com.into.websoso.core.resource.R.string.inquire_link
+import com.into.websoso.core.resource.R.string.novel_detail_feed
+import com.into.websoso.core.resource.R.string.novel_detail_feed_write
+import com.into.websoso.core.resource.R.string.novel_detail_info
+import com.into.websoso.core.resource.R.string.novel_detail_remove_accept
+import com.into.websoso.core.resource.R.string.novel_detail_remove_cancel
+import com.into.websoso.core.resource.R.string.novel_detail_remove_evaluate_alert_message
+import com.into.websoso.core.resource.R.string.novel_detail_remove_evaluate_alert_title
+import com.into.websoso.core.resource.R.string.novel_detail_remove_result
 import com.into.websoso.databinding.ActivityNovelDetailBinding
 import com.into.websoso.databinding.ItemNovelDetailTooltipBinding
 import com.into.websoso.databinding.MenuNovelDetailPopupBinding
-import com.into.websoso.resource.R.drawable.ic_novel_detail_check
-import com.into.websoso.resource.R.string.inquire_link
-import com.into.websoso.resource.R.string.novel_detail_feed
-import com.into.websoso.resource.R.string.novel_detail_feed_write
-import com.into.websoso.resource.R.string.novel_detail_info
-import com.into.websoso.resource.R.string.novel_detail_remove_accept
-import com.into.websoso.resource.R.string.novel_detail_remove_cancel
-import com.into.websoso.resource.R.string.novel_detail_remove_evaluate_alert_message
-import com.into.websoso.resource.R.string.novel_detail_remove_evaluate_alert_title
-import com.into.websoso.resource.R.string.novel_detail_remove_result
 import com.into.websoso.ui.common.dialog.LoginRequestDialogFragment
 import com.into.websoso.ui.createFeed.CreateFeedActivity
 import com.into.websoso.ui.feedDetail.model.EditFeedModel
