+++ app/src/main/java/com/into/websoso/ui/novelFeed/NovelFeedFragment.kt
@@ -25,16 +25,16 @@ import com.into.websoso.core.common.util.InfiniteScrollListener
 import com.into.websoso.core.common.util.SingleEventHandler
 import com.into.websoso.core.common.util.showWebsosoSnackBar
 import com.into.websoso.core.common.util.toIntPxFromDp
+import com.into.websoso.core.resource.R.drawable.ic_blocked_user_snack_bar
+import com.into.websoso.core.resource.R.drawable.ic_novel_detail_check
+import com.into.websoso.core.resource.R.string.block_user_success_message
+import com.into.websoso.core.resource.R.string.feed_popup_menu_content_isMyFeed
+import com.into.websoso.core.resource.R.string.feed_popup_menu_content_report_isNotMyFeed
+import com.into.websoso.core.resource.R.string.other_user_page_withdraw_user
 import com.into.websoso.databinding.DialogRemovePopupMenuBinding
 import com.into.websoso.databinding.DialogReportPopupMenuBinding
 import com.into.websoso.databinding.FragmentNovelFeedBinding
 import com.into.websoso.databinding.MenuFeedPopupBinding
-import com.into.websoso.resource.R.drawable.ic_blocked_user_snack_bar
-import com.into.websoso.resource.R.drawable.ic_novel_detail_check
-import com.into.websoso.resource.R.string.block_user_success_message
-import com.into.websoso.resource.R.string.feed_popup_menu_content_isMyFeed
-import com.into.websoso.resource.R.string.feed_popup_menu_content_report_isNotMyFeed
-import com.into.websoso.resource.R.string.other_user_page_withdraw_user
 import com.into.websoso.ui.common.dialog.LoginRequestDialogFragment
 import com.into.websoso.ui.createFeed.CreateFeedActivity
 import com.into.websoso.ui.feedDetail.FeedDetailActivity
