+++ app/src/main/java/com/into/websoso/ui/novelInfo/NovelInfoFragment.kt
@@ -19,13 +19,13 @@ import com.into.websoso.R.style.body2
 import com.into.websoso.core.common.ui.base.BaseFragment
 import com.into.websoso.core.common.ui.custom.WebsosoChip
 import com.into.websoso.core.common.util.tracker.Tracker
+import com.into.websoso.core.resource.R.string.novel_info_charm_points_body
+import com.into.websoso.core.resource.R.string.novel_info_keyword_chip_text
+import com.into.websoso.core.resource.R.string.novel_info_read_status_quit_count
+import com.into.websoso.core.resource.R.string.novel_info_read_status_watched_count
+import com.into.websoso.core.resource.R.string.novel_info_read_status_watching_count
+import com.into.websoso.core.resource.R.string.novel_info_user_unit
 import com.into.websoso.databinding.FragmentNovelInfoBinding
-import com.into.websoso.resource.R.string.novel_info_charm_points_body
-import com.into.websoso.resource.R.string.novel_info_keyword_chip_text
-import com.into.websoso.resource.R.string.novel_info_read_status_quit_count
-import com.into.websoso.resource.R.string.novel_info_read_status_watched_count
-import com.into.websoso.resource.R.string.novel_info_read_status_watching_count
-import com.into.websoso.resource.R.string.novel_info_user_unit
 import com.into.websoso.ui.novelInfo.component.NovelInfoPlatformsContainer
 import com.into.websoso.ui.novelInfo.model.ExpandTextUiModel
 import com.into.websoso.ui.novelInfo.model.KeywordModel
