+++ app/src/main/java/com/into/websoso/ui/novelRating/NovelRatingActivity.kt
@@ -25,17 +25,17 @@ import com.into.websoso.core.common.util.showWebsosoSnackBar
 import com.into.websoso.core.common.util.showWebsosoToast
 import com.into.websoso.core.common.util.toFloatPxFromDp
 import com.into.websoso.core.common.util.tracker.Tracker
+import com.into.websoso.core.resource.R.drawable.ic_novel_detail_check
+import com.into.websoso.core.resource.R.drawable.ic_novel_rating_alert
+import com.into.websoso.core.resource.R.drawable.ic_novel_rating_keword_remove
+import com.into.websoso.core.resource.R.string.novel_rating_cancel_alert_accept
+import com.into.websoso.core.resource.R.string.novel_rating_cancel_alert_cancel
+import com.into.websoso.core.resource.R.string.novel_rating_cancel_alert_title
+import com.into.websoso.core.resource.R.string.novel_rating_charm_point_exceed
+import com.into.websoso.core.resource.R.string.novel_rating_charm_points
+import com.into.websoso.core.resource.R.string.novel_rating_complete
+import com.into.websoso.core.resource.R.string.novel_rating_save_error
 import com.into.websoso.databinding.ActivityNovelRatingBinding
-import com.into.websoso.resource.R.drawable.ic_novel_detail_check
-import com.into.websoso.resource.R.drawable.ic_novel_rating_alert
-import com.into.websoso.resource.R.drawable.ic_novel_rating_keword_remove
-import com.into.websoso.resource.R.string.novel_rating_cancel_alert_accept
-import com.into.websoso.resource.R.string.novel_rating_cancel_alert_cancel
-import com.into.websoso.resource.R.string.novel_rating_cancel_alert_title
-import com.into.websoso.resource.R.string.novel_rating_charm_point_exceed
-import com.into.websoso.resource.R.string.novel_rating_charm_points
-import com.into.websoso.resource.R.string.novel_rating_complete
-import com.into.websoso.resource.R.string.novel_rating_save_error
 import com.into.websoso.ui.novelDetail.NovelAlertDialogFragment
 import com.into.websoso.ui.novelDetail.model.NovelAlertModel
 import com.into.websoso.ui.novelRating.model.CharmPoint
