+++ app/src/main/java/com/into/websoso/ui/profileEdit/ProfileEditActivity.kt
@@ -32,14 +32,14 @@ import com.into.websoso.core.common.ui.model.ResultFrom.ProfileEditSuccess
 import com.into.websoso.core.common.util.getS3ImageUrl
 import com.into.websoso.core.common.util.showWebsosoToast
 import com.into.websoso.core.common.util.toFloatPxFromDp
+import com.into.websoso.core.resource.R.drawable.ic_novel_detail_check
+import com.into.websoso.core.resource.R.drawable.ic_novel_rating_alert
+import com.into.websoso.core.resource.R.string.novel_rating_save_error
+import com.into.websoso.core.resource.R.string.profile_edit_introduction_max_count
+import com.into.websoso.core.resource.R.string.profile_edit_nickname_max_count
+import com.into.websoso.core.resource.R.string.profile_edit_success
 import com.into.websoso.databinding.ActivityProfileEditBinding
 import com.into.websoso.domain.model.NicknameValidationResult.VALID_NICKNAME
-import com.into.websoso.resource.R.drawable.ic_novel_detail_check
-import com.into.websoso.resource.R.drawable.ic_novel_rating_alert
-import com.into.websoso.resource.R.string.novel_rating_save_error
-import com.into.websoso.resource.R.string.profile_edit_introduction_max_count
-import com.into.websoso.resource.R.string.profile_edit_nickname_max_count
-import com.into.websoso.resource.R.string.profile_edit_success
 import com.into.websoso.ui.profileEdit.model.Genre
 import com.into.websoso.ui.profileEdit.model.Genre.Companion.toGenreFromKr
 import com.into.websoso.ui.profileEdit.model.LoadProfileResult
