+++ app/src/main/java/com/into/websoso/ui/setting/SettingActivity.kt
@@ -14,15 +14,15 @@ import com.into.websoso.R.layout.activity_setting
 import com.into.websoso.core.common.ui.base.BaseActivity
 import com.into.websoso.core.common.ui.model.ResultFrom.ChangeProfileDisclosure
 import com.into.websoso.core.common.util.showWebsosoSnackBar
+import com.into.websoso.core.resource.R.drawable.ic_novel_detail_check
+import com.into.websoso.core.resource.R.string.inquire_link
+import com.into.websoso.core.resource.R.string.privacy_policy_link
+import com.into.websoso.core.resource.R.string.profile_disclosure_message
+import com.into.websoso.core.resource.R.string.profile_disclosure_private
+import com.into.websoso.core.resource.R.string.profile_disclosure_public
+import com.into.websoso.core.resource.R.string.terms_of_use_link
+import com.into.websoso.core.resource.R.string.websoso_official
 import com.into.websoso.databinding.ActivitySettingBinding
-import com.into.websoso.resource.R.drawable.ic_novel_detail_check
-import com.into.websoso.resource.R.string.inquire_link
-import com.into.websoso.resource.R.string.privacy_policy_link
-import com.into.websoso.resource.R.string.profile_disclosure_message
-import com.into.websoso.resource.R.string.profile_disclosure_private
-import com.into.websoso.resource.R.string.profile_disclosure_public
-import com.into.websoso.resource.R.string.terms_of_use_link
-import com.into.websoso.resource.R.string.websoso_official
 import com.into.websoso.ui.accountInfo.AccountInfoActivity
 import com.into.websoso.ui.notificationSetting.NotificationSettingActivity
 import com.into.websoso.ui.profileDisclosure.ProfileDisclosureActivity
